feat: add task_manager utils
